bot/requirements.txt
# Telegram Bot Dependencies
python-telegram-bot[job-queue]>=21.0
supabase>=2.10.0
openai>=1.12.0
python-dotenv>=1.0.1
pytz>=2024.1
